requirements.txt
dnspython==2.6.1
pillow==10.3.0
pymongo==4.7.3
python-dotenv==1.0.1
ttkbootstrap==1.10.1
